+++ hash/src/sha3_arm_ce.c
@@ -125,7 +125,7 @@ static void keccak_arm_ce_permutation(void *state, unsigned int rounds)
 }
 
 static void keccak_arm_ce_add_bytes(void *state, const unsigned char *data,
-				    unsigned int offset, unsigned int length)
+				    size_t offset, size_t length)
 {
 	uint8_t *_state = (uint8_t *)state;
